refactor: 重构keymap，添加平台判断，抽取成独立工具
