+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+export function isMac(): boolean {
+  return /mac/i.test(navigator.userAgent);
+}
+export function isWin(): boolean {
+  return /windows/i.test(navigator.userAgent);
+}
